+++ tests/apps/demultiplex/conftest.py
@@ -13,23 +13,11 @@ from cg.apps.lims.samplesheet import (
 from cg.models.demultiplex.run_parameters import RunParameters
 
 
-@pytest.fixture(name="demultiplexed_runs_dir")
-def fixture_demultiplexed_runs_dir(demultiplex_fixtures: Path) -> Path:
-    """Return the path to a dir with flowcells ready for demultiplexing"""
-    return Path(demultiplex_fixtures, "demultiplexed-runs")
-
-
 @pytest.fixture(name="output_dirs_bcl2fastq")
-def fixture_output_dirs_bcl2fastq(demultiplexed_runs_dir: Path) -> Path:
-    """Return the output path a dir with flowcells that have finished demultiplexing using
-    bcl2fastq"""
-    return Path(demultiplexed_runs_dir, "bcl2fastq")
-
-
-@pytest.fixture(name="demux_run_dir")
-def fixture_demux_run_dir(demultiplex_fixtures: Path) -> Path:
-    """Return the path to a dir with flowcells ready for demultiplexing"""
-    return Path(demultiplex_fixtures, "flowcell-runs")
+def fixture_output_dirs_bcl2fastq(demultiplexed_runs: Path) -> Path:
+    """Return the output path a dir with flow cells that have finished demultiplexing using
+    bcl2fastq."""
+    return Path(demultiplexed_runs, "bcl2fastq")
 
 
 @pytest.fixture(name="demux_run_dir_bcl2fastq")
