+++ tests/models/demultiplexing/test_flowcell_model.py
@@ -5,6 +5,7 @@ from cg.models.demultiplex.flowcell import Flowcell
 
 
 def test_flowcell_id(flowcell_path: Path):
+    """Test parsing of flow cell id."""
     # GIVEN the path to a finished flow cell run
     # GIVEN the flow cell id
     flowcell_id: str = flowcell_path.name.split("_")[-1][1:]
@@ -17,6 +18,7 @@ def test_flowcell_id(flowcell_path: Path):
 
 
 def test_flowcell_position(flowcell_path: Path):
+    """Test getting flow cell position."""
     # GIVEN the path to a finished flow cell
     # GIVEN a flow cell object
     flowcell_obj = Flowcell(flowcell_path)
@@ -29,6 +31,7 @@ def test_flowcell_position(flowcell_path: Path):
 
 
 def test_rta_exists(flowcell_object: Flowcell):
+    """Test return of RTS file."""
     # GIVEN the path to a finished flow cell
     # GIVEN a flow cell object
 
@@ -51,16 +54,18 @@ def test_is_hiseq_x_copy_completed_ready(flowcell_object: Flowcell):
     assert is_completed is True
 
 
-def test_is_hiseq_x_delivery_started_ready(flowcell_object: Flowcell):
+def test_is_hiseq_x_delivery_started_ready(
+    flowcell_object: Flowcell, demultiplexing_delivery_file: Path
+):
     # GIVEN the path to a demultiplexed finished flow cell
     # GIVEN a flow cell object
     # GIVEN a delivery file
-    Path(flowcell_object.path, DemultiplexingDirsAndFiles.DELIVERY).touch()
+    demultiplexing_delivery_file.touch()
 
     # WHEN checking the path to the delivery file
     is_delivered = flowcell_object.is_hiseq_x_delivery_started()
 
-    Path(flowcell_object.path, DemultiplexingDirsAndFiles.DELIVERY).unlink()
+    demultiplexing_delivery_file.unlink()
 
     # THEN assert that the file exists
     assert is_delivered is True
@@ -77,17 +82,17 @@ def test_is_hiseq_x_delivery_started_not_ready(flowcell_object: Flowcell):
     assert is_delivered is False
 
 
-def test_is_hiseq_x(flowcell_object: Flowcell):
+def test_is_hiseq_x(flowcell_object: Flowcell, hiseq_x_tile_dir: Path):
     # GIVEN the path to a demultiplexed finished flow cell
     # GIVEN a flow cell object
     # GIVEN a Hiseq X directory
-    Path(flowcell_object.path, DemultiplexingDirsAndFiles.HiseqX_TILE_DIR).mkdir(exist_ok=True)
+    hiseq_x_tile_dir.mkdir(exist_ok=True)
 
     # WHEN checking the path to the Hiseq X flow cell directory
     is_hiseq_x = flowcell_object.is_hiseq_x()
 
     # Clean up
-    Path(flowcell_object.path, DemultiplexingDirsAndFiles.HiseqX_TILE_DIR).rmdir()
+    hiseq_x_tile_dir.rmdir()
 
     # THEN assert that the file exists
     assert is_hiseq_x is True
